Move container initialization to inside `DrawableNote`
