+++ maisim/maisim.Game/Component/Gameplay/Notes/DrawableNote.cs
@@ -1,4 +1,9 @@
-﻿using osu.Framework.Graphics.Containers;
+﻿using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Sprites;
+using osu.Framework.Graphics.Textures;
+using osuTK;
 
 namespace maisim.Game.Component.Gameplay.Notes
 {
@@ -7,6 +12,20 @@ namespace maisim.Game.Component.Gameplay.Notes
     /// </summary>
     public abstract class DrawableNote : CompositeDrawable
     {
+        private Drawable[] noteParts;
 
+        [BackgroundDependencyLoader]
+        private void load(TextureStore textureStore)
+        {
+            InternalChild = new Container
+            {
+                Anchor = Anchor.Centre,
+                Origin = Anchor.Centre,
+                Size = new Vector2(50),
+                Children = AddNoteParts(textureStore)
+            };
+        }
+
+        public abstract Drawable[] AddNoteParts(TextureStore textureStore);
     }
 }
