+++ maisim/maisim.Game/Component/Gameplay/Notes/DrawableSlidePathNote.cs
@@ -12,20 +12,17 @@ namespace maisim.Game.Component.Gameplay.Notes
     /// </summary>
     public class DrawableSlidePathNote : DrawableNote
     {
-        [BackgroundDependencyLoader]
-        private void load(TextureStore textureStore)
+        public override Drawable[] AddNoteParts(TextureStore textureStore)
         {
-            InternalChild = new Container
+            return new Drawable[]
             {
-                Anchor = Anchor.Centre,
-                Origin = Anchor.Centre,
-                Size = new Vector2(20),
-                Child = new Sprite
+                new Sprite
                 {
                     RelativeSizeAxes = Axes.Both,
                     Anchor = Anchor.Centre,
                     Origin = Anchor.Centre,
                     FillMode = FillMode.Fill,
+                    Scale = new Vector2(0.5f),
                     Texture = textureStore.Get("Notes/SlidePath.png")
                 }
             };
