+++ maisim/maisim.Game/Component/Gameplay/Notes/DrawableTapNote.cs
@@ -12,15 +12,11 @@ namespace maisim.Game.Component.Gameplay.Notes
     /// </summary>
     public class DrawableTapNote : DrawableNote
     {
-        [BackgroundDependencyLoader]
-        private void load(TextureStore textureStore)
+        public override Drawable[] AddNoteParts(TextureStore textureStore)
         {
-            InternalChild = new Container
+            return new Drawable[]
             {
-                Anchor = Anchor.Centre,
-                Origin = Anchor.Centre,
-                Size = new Vector2(50),
-                Child = new Sprite
+                new Sprite
                 {
                     RelativeSizeAxes = Axes.Both,
                     Anchor = Anchor.Centre,
