+++ maisim/maisim.Game/Component/Gameplay/Notes/DrawableTouchHoldNote.cs
@@ -7,20 +7,16 @@ using osuTK;
 
 namespace maisim.Game.Component.Gameplay.Notes
 {
+    /// <summary>
+    /// Class represent the TOUCH HOLD note.
+    /// </summary>
     public class TouchHold : DrawableNote
     {
-        /// <summary>
-        /// Class represent the TOUCH HOLD note.
-        /// </summary>
-        [BackgroundDependencyLoader]
-        private void load(TextureStore textureStore)
+        public override Drawable[] AddNoteParts(TextureStore textureStore)
         {
-            InternalChild = new Container
+            return new Drawable[]
             {
-                Anchor = Anchor.Centre,
-                Origin = Anchor.Centre,
-                Size = new Vector2(50),
-                Child = new Sprite
+                new Sprite
                 {
                     RelativeSizeAxes = Axes.Both,
                     Anchor = Anchor.Centre,
