+++ maisim/maisim.Game/Component/Gameplay/Notes/DrawableTouchNote.cs
@@ -12,64 +12,57 @@ namespace maisim.Game.Component.Gameplay.Notes
     /// </summary>
     public class DrawableTouchNote : DrawableNote
     {
-        [BackgroundDependencyLoader]
-        private void load(TextureStore textureStore)
+        public override Drawable[] AddNoteParts(TextureStore textureStore)
         {
-            InternalChild = new Container
+            return new Drawable[]
             {
-                Anchor = Anchor.Centre,
-                Origin = Anchor.Centre,
-                Size = new Vector2(50),
-                Children = new Drawable[]
+                new Sprite
                 {
-                    new Sprite
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        FillMode = FillMode.Fill,
-                        Texture = textureStore.Get("Notes/TouchDot.png")
-                    },
-                    new Sprite
-                    {
-                        Name = "UpArrow",
-                        Position = new Vector2(0, -20),
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Rotation = 0,
-                        Size = new Vector2(30, 15),
-                        Texture = textureStore.Get("Notes/TouchArrow.png")
-                    },
-                    new Sprite
-                    {
-                        Name = "RightArrow",
-                        Position = new Vector2(20, 0),
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Rotation = 90,
-                        Size = new Vector2(30, 15),
-                        Texture = textureStore.Get("Notes/TouchArrow.png")
-                    },
-                    new Sprite
-                    {
-                        Name = "DownArrow",
-                        Position = new Vector2(0, 20),
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Rotation = 180,
-                        Size = new Vector2(30, 15),
-                        Texture = textureStore.Get("Notes/TouchArrow.png")
-                    },
-                    new Sprite
-                    {
-                        Name = "LeftArrow",
-                        Position = new Vector2(-20, 0),
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Rotation = 270,
-                        Size = new Vector2(30, 15),
-                        Texture = textureStore.Get("Notes/TouchArrow.png")
-                    }
+                    RelativeSizeAxes = Axes.Both,
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    FillMode = FillMode.Fill,
+                    Texture = textureStore.Get("Notes/TouchDot.png")
+                },
+                new Sprite
+                {
+                    Name = "UpArrow",
+                    Position = new Vector2(0, -20),
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Rotation = 0,
+                    Size = new Vector2(30, 15),
+                    Texture = textureStore.Get("Notes/TouchArrow.png")
+                },
+                new Sprite
+                {
+                    Name = "RightArrow",
+                    Position = new Vector2(20, 0),
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Rotation = 90,
+                    Size = new Vector2(30, 15),
+                    Texture = textureStore.Get("Notes/TouchArrow.png")
+                },
+                new Sprite
+                {
+                    Name = "DownArrow",
+                    Position = new Vector2(0, 20),
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Rotation = 180,
+                    Size = new Vector2(30, 15),
+                    Texture = textureStore.Get("Notes/TouchArrow.png")
+                },
+                new Sprite
+                {
+                    Name = "LeftArrow",
+                    Position = new Vector2(-20, 0),
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Rotation = 270,
+                    Size = new Vector2(30, 15),
+                    Texture = textureStore.Get("Notes/TouchArrow.png")
                 }
             };
         }
